+++ src/Markdown.php
@@ -15,6 +15,8 @@ namespace Eventum;
 
 use Eventum\CommonMark\MentionExtension;
 use Eventum\EventDispatcher\EventManager;
+use HTMLPurifier;
+use HTMLPurifier_HTML5Config;
 use League\CommonMark\CommonMarkConverter;
 use League\CommonMark\ConverterInterface;
 use League\CommonMark\Environment;
@@ -36,8 +38,15 @@ class Markdown
      */
     private const MAX_NESTING_LEVEL = 500;
 
+    /** @var HTMLPurifier */
+    private $purifier;
     /** @var ConverterInterface[] */
     private $converter = [];
+
+    public function __construct()
+    {
+        $this->purifier = $this->createPurifier();
+    }
 
     public function render(string $text): string
     {
@@ -45,7 +54,10 @@ class Markdown
             return $text;
         }
 
-        return $this->getConverter(false)->convertToHtml($text);
+        $html = $this->getConverter(false)->convertToHtml($text);
+        $html = $this->getPurifier()->purify($html);
+
+        return $html;
     }
 
     public function renderInline(string $text): string
@@ -54,12 +66,20 @@ class Markdown
             return $text;
         }
 
-        return $this->getConverter(true)->convertToHtml($text);
+        $html = $this->getConverter(true)->convertToHtml($text);
+        $html = $this->getPurifier()->purify($html);
+
+        return $html;
     }
 
     private function getConverter(bool $inline): ConverterInterface
     {
-        return $this->converter[(int)$inline] ?? $this->createConverter($inline);
+        return $this->converter[(int)$inline] ?? $this->converter[(int)$inline] = $this->createConverter($inline);
+    }
+
+    private function getPurifier(): HTMLPurifier
+    {
+        return $this->purifier ?? $this->purifier = $this->createPurifier();
     }
 
     private function createConverter(bool $inline): ConverterInterface
@@ -85,6 +105,36 @@ class Markdown
         $this->applyExtensions($environment);
 
         return new CommonMarkConverter([], $environment);
+    }
+
+    private function createPurifier(): HTMLPurifier
+    {
+        $cacheDir = APP_VAR_PATH . '/cache/purifier';
+        is_dir($cacheDir) || mkdir($cacheDir, 02775) || is_dir($cacheDir);
+
+        $config = HTMLPurifier_HTML5Config::createDefault();
+
+        $config->set('AutoFormat.AutoParagraph', true);
+        // remove empty tag pairs
+        $config->set('AutoFormat.RemoveEmpty', true);
+        // remove empty, even if it contains an &nbsp;
+        $config->set('AutoFormat.RemoveEmpty.RemoveNbsp', true);
+        // preserve html comments
+        $config->set('HTML.AllowedCommentsRegexp', '/.+/');
+
+        // disable tidy processing, even if extension present
+        $config->set('Output.TidyFormat', false);
+
+        // disable useless normalizer we do not need
+        $config->set('Core.NormalizeNewlines', false);
+
+        // allow tasklist <input> checkboxes
+        $config->set('HTML.Trusted', true);
+
+        // Absolute path with no trailing slash to store serialized definitions in.
+        $config->set('Cache.SerializerPath', $cacheDir);
+
+        return new HTMLPurifier($config);
     }
 
     private function applyExtensions(Environment $environment): void